Move all secondary indexes interfaces into its own file
